[eas-cli] extract app lookup params
