+++ packages/eas-cli/src/credentials/ios/actions/AssignAscApiKey.ts
@@ -4,7 +4,7 @@ import {
 } from '../../../graphql/generated';
 import Log from '../../../log';
 import { CredentialsContext } from '../../context';
-import { AppLookupParams } from '../api/GraphqlClient';
+import { AppLookupParams } from '../api/graphql/types/AppLookupParams';
 import { resolveAppleTeamIfAuthenticatedAsync } from './AppleTeamUtils';
 import { AppStoreApiKeyPurpose } from './AscApiKeyUtils';
