+++ packages/eas-cli/src/credentials/ios/actions/ConfigureProvisioningProfile.ts
@@ -10,8 +10,8 @@ import { ora } from '../../../ora';
 import { getApplePlatformFromSdkRoot } from '../../../project/ios/target';
 import { CredentialsContext } from '../../context';
 import { MissingCredentialsNonInteractiveError } from '../../errors';
-import { AppLookupParams } from '../api/GraphqlClient';
 import { AppleProvisioningProfileMutationResult } from '../api/graphql/mutations/AppleProvisioningProfileMutation';
+import { AppLookupParams } from '../api/graphql/types/AppLookupParams';
 import { ProvisioningProfileStoreInfo } from '../appstore/Credentials.types';
 import { AuthCtx } from '../appstore/authenticateTypes';
 import { Target } from '../types';
