+++ packages/eas-cli/src/credentials/ios/actions/SetUpDistributionCertificate.ts
@@ -10,8 +10,8 @@ import { confirmAsync, promptAsync } from '../../../prompts';
 import sortBy from '../../../utils/expodash/sortBy';
 import { CredentialsContext } from '../../context';
 import { MissingCredentialsNonInteractiveError } from '../../errors';
-import { AppLookupParams } from '../api/GraphqlClient';
 import { AppleDistributionCertificateMutationResult } from '../api/graphql/mutations/AppleDistributionCertificateMutation';
+import { AppLookupParams } from '../api/graphql/types/AppLookupParams';
 import { getValidCertSerialNumbers } from '../appstore/CredentialsUtils';
 import { AppleTeamMissingError } from '../errors';
 import { resolveAppleTeamIfAuthenticatedAsync } from './AppleTeamUtils';
