+++ packages/eas-cli/src/credentials/ios/api/graphql/types/AppLookupParams.ts
@@ -1,0 +1,8 @@
+import { Account } from '../../../../../user/Account';
+
+export interface AppLookupParams {
+  account: Account;
+  projectName: string;
+  bundleIdentifier: string;
+  parentBundleIdentifier?: string;
+}
