+++ packages/eas-cli/src/credentials/ios/utils/printCredentials.ts
@@ -12,7 +12,7 @@ import { APPLE_DEVICE_CLASS_LABELS } from '../../../graphql/types/credentials/Ap
 import Log from '../../../log';
 import { fromNow } from '../../../utils/date';
 import formatFields from '../../../utils/formatFields';
-import { AppLookupParams } from '../api/GraphqlClient';
+import { AppLookupParams } from '../api/graphql/types/AppLookupParams';
 import { App, IosAppBuildCredentialsMap, IosAppCredentialsMap, Target } from '../types';
 
 function prettyIosDistributionType(distributionType: IosDistributionType): string {
